Uac 55 1

+++ api/db_info/crud.py
@@ -20,10 +20,21 @@ def create_item(db: Session, new_item: schemas.ItemCreate):
     db.refresh(db_item)
     return db_item
 
-def delete_item(db: Session, item: schemas.Item):
-    db_item = db.query(models.Item).filter(models.Item.id == item.id).first()
+def retrieve_item(db: Session, id: int):
+    db_item = get_item_by_id(db, id)
+    if db_item == None:
+        return None
+    if db_item.dropoffPoint_id != None and db_item.mail == None:
+        db_item.state = "retrieved"
+        db_item.dropoffPoint_id = None
+        db.flush()
+        db.commit()
+    return db_item
+
+def delete_item(db: Session, id: int):
+    db_item = get_item_by_id(db, id)
+    if db_item == None:
+        return None
     db.delete(db_item)
     db.commit()
-    if get_item_by_id(db, item.id):
-        return "ERROR"
     return "OK"
